feat: (#3.1): Transfer model and logic to system C

+++ Route256/ApiModels/V1/GetHistory/GetHistoryResponse.cs
@@ -3,4 +3,4 @@ namespace Route256.ApiModels.V1.GetHistory;
 public record GetHistoryResponse(CargoHistoryResponse[] CargoHistory);
 
 
-public record CargoHistoryResponse(decimal Volume, decimal Price);
+public record CargoHistoryResponse(double Volume, decimal Price);
