+++ Route256/ApiModels/V3/GetHistory/GetHistoryResponse.cs
@@ -3,4 +3,4 @@ namespace Route256.ApiModels.V3.GetHistory;
 public record GetHistoryResponse(CargoHistoryResponse[] CargoHistory);
 
 
-public record CargoHistoryResponse(decimal Volume, decimal Weight, decimal Price, int Distance);
+public record CargoHistoryResponse(double Volume, double Weight, decimal Price, int Distance);
