+++ Route256/BLL/DeliveryPrice/Models/Cargo.cs
@@ -2,7 +2,7 @@ namespace Route256.BLL.DeliveryPrice.Models;
 
 public class Cargo
 {
-    public Cargo(decimal volume, decimal weight, decimal price, int? distance = null)
+    public Cargo(double volume, double weight, decimal price, int? distance = null)
     {
         Volume = volume;
         Weight = weight;
@@ -10,9 +10,9 @@ public class Cargo
         Distance = distance;
     }
 
-    public decimal Volume { get; set; }
+    public double Volume { get; set; }
     
-    public decimal Weight { get; set; }
+    public double Weight { get; set; }
     
     public decimal Price { get; set; }
