+++ Route256/DAL/Entities/CargoDb.cs
@@ -2,9 +2,9 @@ namespace Route256.DAL.Entities;
 
 public class CargoDb
 {
-    public decimal Volume { get; set; }
+    public double Volume { get; set; }
     
-    public decimal Weight { get; set; }
+    public double Weight { get; set; }
     
     public decimal Price { get; set; }
